Split test_model_version
